% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin_ts.R
\name{coin_ts}
\alias{coin_ts}
\title{Fetch time series data of two cryptocurrency exchange rates.}
\usage{
coin_ts(from = "BTC", to = "USDT", start = "2001-01-01",
  end = "9999-01-01", period = 86400)
}
\value{
a tibble with data of HLOCV of the referred cryptocoin
}
\description{
Fetch time series data of two cryptocurrency exchange rates.
}
\examples{
btc_data <- coin_ts(from="BTC", to="USDT", start="2001-01-01", end="2017-02-18")
}

